Display error when missing currency; fixed json validation

+++ app/Services/OpenExchangeCurrencyApiService.php
@@ -37,13 +37,19 @@ class OpenExchangeCurrencyApiService implements CurrencyApiInterface
         if ($response->successful()) {
 
             $insert = [];
+            $usd = false;
+            $eur = false;
             $allCurrencies = $response->json();
             foreach ($allCurrencies as $shortcut => $fullName) {
                 $insert[] = ['symbol' => $shortcut, 'name' => $fullName, 'combined_name' => $shortcut . ' - ' . $fullName];
+                $shortcut === 'USD' ? $usd = true : false;
+                $shortcut === 'EUR' ? $eur = true : false;
             }
-
-            $this->currencies->truncate();
-            $this->currencies->insert($insert);
+            if (count($insert) && $usd && $eur) { //check if array has any rows and some major currencies are present
+                $this->currencies->truncate();
+                $this->currencies->insert($insert);
+            }
+            
         }
     }
